+++ jsonpath4k/src/commonTest/kotlin/at/asitplus/jsonpath/DependencyManagementTest.kt
@@ -3,7 +3,6 @@ package at.asitplus.jsonpath
 import at.asitplus.jsonpath.core.JsonPathCompiler
 import at.asitplus.jsonpath.core.JsonPathFunctionExtension
 import at.asitplus.jsonpath.core.JsonPathQuery
-import at.asitplus.jsonpath.core.JsonPathSelector
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.collections.shouldBeIn
@@ -15,20 +14,21 @@ import kotlinx.serialization.json.buildJsonObject
 @Suppress("unused")
 class DependencyManagementTest : FreeSpec({
     // making sure that the dependencies are reset to their default for the next test
-    val defaultCompilerBuilderBackup = JsonPathDependencyManager.compiler
+    val defaultCompilerBuilderBackup = JsonPath.defaultCompiler
     val defaultFunctionExtensionRepositoryBackup =
-        JsonPathDependencyManager.functionExtensionRepository.export()
+        JsonPath.defaultFunctionExtensionRepository.export()
+
     beforeEach {
         // prepare a dummy repository to be modified by the tests
-        JsonPathDependencyManager.functionExtensionRepository =
+        JsonPath.defaultFunctionExtensionRepository =
             JsonPathFunctionExtensionMapRepository(
                 defaultFunctionExtensionRepositoryBackup.toMutableMap()
             )
     }
     afterEach {
-        JsonPathDependencyManager.apply {
-            compiler = defaultCompilerBuilderBackup
-            functionExtensionRepository = JsonPathFunctionExtensionMapRepository(
+        JsonPath.Companion.apply {
+            defaultCompiler = defaultCompilerBuilderBackup
+            defaultFunctionExtensionRepository = JsonPathFunctionExtensionMapRepository(
                 defaultFunctionExtensionRepositoryBackup.toMutableMap()
             )
         }
@@ -39,7 +39,7 @@ class DependencyManagementTest : FreeSpec({
             val jsonPathStatement = "$[?foo()]"
 
             val jsonPath = shouldNotThrowAny {
-                val testRepo = JsonPathDependencyManager.functionExtensionRepository.export().plus(
+                val testRepo = JsonPath.defaultFunctionExtensionRepository.export().plus(
                     "foo" to JsonPathFunctionExtension.LogicalTypeFunctionExtension {
                         true
                     }
@@ -68,7 +68,7 @@ class DependencyManagementTest : FreeSpec({
                 selectors = listOf()
             )
         }
-        JsonPathDependencyManager.compiler = incorrectEmptyQueryCompiler
+        JsonPath.defaultCompiler = incorrectEmptyQueryCompiler
         val emptyQueryResult = JsonPath("$").query(buildJsonObject {})
         // this checks, whether the new compiler has indeed been used
         emptyQueryResult.shouldHaveSize(0)